+++ packages/terra-avatar/tests/jest/Avatar.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Avatar from '../../src/Avatar';
-import exampleProfilePhoto from '../assets/150x150.jpg';
+import exampleProfilePhoto from '../../examples/assets/150x150.jpg';
 
 describe('Avatar', () => {
   // Snapshot Tests