+++ packages/terra-doc-template/src/ExampleTemplate.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ToggleButton from 'terra-toggle-button';
-import SyntaxHighlighter from 'react-syntax-highlighter/prism';
+import SyntaxHighlighter, { registerLanguage } from 'react-syntax-highlighter/prism-light';
 import { okaidia } from 'react-syntax-highlighter/styles/prism';
+import jsx from 'react-syntax-highlighter/languages/prism/jsx';
+
+registerLanguage('jsx', jsx);
 
 const propTypes = {
   /**
@@ -20,7 +23,7 @@ const propTypes = {
   /**
    * An optional description that will be displayed below the title as regular text
    */
-  description: PropTypes.string,
+  description: PropTypes.node,
   /**
    * The children components for this example which will be displayed below the example
    */
@@ -49,11 +52,14 @@ class IndexExampleTemplate extends React.Component {
   render() {
     const { title, example, exampleSrc, children, description, ...customProps } = this.props;
 
-    const indexExampleSrc = (
-      <ToggleButton isAnimated closedButtonText="View Source Code" data-terra-source-code-toggle>
-        <SyntaxHighlighter language="javascript" style={okaidia}>{exampleSrc}</SyntaxHighlighter>
-      </ToggleButton>
-    );
+    let indexExampleSrc;
+    if (exampleSrc) {
+      indexExampleSrc = (
+        <ToggleButton isAnimated closedButtonText="View Source Code" data-terra-source-code-toggle>
+          <SyntaxHighlighter language="javascript" style={okaidia}>{exampleSrc}</SyntaxHighlighter>
+        </ToggleButton>
+      );
+    }
 
     return (
       <div {...customProps}>