+++ packages/terra-site/src/examples/image/Index.jsx
@@ -1,23 +1,35 @@
+/* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
-import PropsTable from 'terra-props-table';
-import Markdown from 'terra-markdown';
+import DocTemplate from 'terra-doc-template';
 import ReadMe from 'terra-image/docs/README.md';
-import { version } from 'terra-image/package.json';
+import { name } from 'terra-image/package.json';
 
 // Component Source
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
 import ImageSrc from '!raw-loader!terra-image/src/Image';
 
 // Example Files
 import ImageDefault from './ImageDefault';
+import ImageDefaultSrc from '!raw-loader!./ImageDefault';
 
-const ImageExamples = () => (
-  <div>
-    <div id="version">Version: {version}</div>
-    <Markdown id="readme" src={ReadMe} />
-    <PropsTable id="props" src={ImageSrc} />
-    <ImageDefault />
-  </div>
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
+    examples={[
+      {
+        title: 'Default Image',
+        example: <ImageDefault />,
+        source: ImageDefaultSrc,
+      },
+    ]}
+    propsTables={[
+      {
+        componentName: 'Image',
+        componentSrc: ImageSrc,
+      },
+    ]}
+  />
 );
 
-export default ImageExamples;
+export default DocPage;