+++ packages/terra-action-header/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Alignment bug with missing start content and title, when using an end aligned item.
+
+1.3.0 - (May 9, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 
 1.2.0 - (May 2, 2018)
 ------------------