+++ packages/terra-doc-template/CHANGELOG.md
@@ -4,7 +4,18 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
+* Add 15px padding to template
+* Limit syntax hilighting support to jsx only to limit package size.
+* Open up description prop to accept node
+
+### Fixed
+* Fix example template to only display example source code toggle when source code is provided
+
+1.1.0 - (May 9, 2018)
+------------------
+### Changed
 * Add syntax highlighting to source code examples
+* Add a link to the component's source code
 
 ### Fixed
 * Update hash key reference in render to match the defined key in the propsTables prop