+++ packages/terra-signature/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Bugfix signature listener not correct after device rotation
+
+1.8.0 - (May 9, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 
 1.7.0 - (May 2, 2018)
 ------------------