+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.0.0 - (May 9, 2018)
+------------------
+### Major Change
+* Updated to use terra-menu v3.0.0
+
 2.8.0 - (May 2, 2018)
 ------------------
 ### Changed